Little cleaner and prevented hard bug, wich freezed Char (Description)
The Bug made the character always swap between  the starting of the hurt-Animation, andd the begin of the Attack Animation.
When the Char was hurt shortly before starting an attack, the attack got not canceled, but couldnt play normal, because the 2 Intervalls where always starting theyre own animation.

So 'gotHurt' could never be resert, and always started as new animation.
And when the attack-Intervall ticked again, it setted the Attack Animation new.
Was really hard to find out.